Custom Exceptions and reponse objects

+++ LM.Application/Features/LeaveAllocation/Handlers/Commands/DeleteLeaveAllocationCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using LM.Application.Features.LeaveAllocation.Requests.Commands;
+﻿using LM.Application.Exceptions;
+using LM.Application.Features.LeaveAllocation.Requests.Commands;
 using LM.Application.Persistence.Contracts;
 using MediatR;
 
@@ -16,6 +17,11 @@ namespace LM.Application.Features.LeaveAllocation.Handlers.Commands
         public async Task<Unit> Handle(DeleteLeaveAllocationCommand request, CancellationToken cancellationToken)
         {
             var leaveAllocation = await _leaveAllocationRepository.Get(request.Id);
+
+            if (leaveAllocation == null)
+            {
+                throw new NotFoundException(nameof(LM.Domain.LeaveAllocation), request.Id);
+            }
 
             await _leaveAllocationRepository.Delete(leaveAllocation);
